requirements.txt
arrow==1.2.3
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
cookiejar==0.0.3
exceptiongroup==1.0.4
Flask==2.2.2
Flask-WTF==1.0.1
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
lxml==4.9.1
MarkupSafe==2.1.1
mechanize==0.4.8
outcome==1.2.0
packaging==21.3
pager==3.3
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
PyYAML==6.0
requests==2.28.1
selenium==4.7.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
text-unidecode==1.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.5
webencodings==0.5.1
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
